hackdavis/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
duckduckgo_search==8.0.1
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.0
greenlet==3.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain-core==0.3.54
langchain-openai==0.3.14
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.32
letta-client==0.1.108
lxml==5.3.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
networkx==3.4.2
numpy==2.2.4
openai==1.75.0
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
postgrest==1.0.1
primp==0.15.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
storage3==0.11.3
streamlit==1.44.1
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.1
syrupy==4.9.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==12.0
yarl==1.19.0
zstandard==0.23.0
